tests_root/tests/requirements.txt
Appium-Python-Client==0.26
astroid==1.6.3
attrs==18.1.0
backports.functools-lru-cache==1.5
configparser==3.5.0
enum34==1.1.6
Faker==0.8.15
funcsigs==1.0.2
futures==3.2.0
glob2==0.6
ipaddress==1.0.22
isort==4.3.4
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.1.0
parse==1.8.2
parse-type==0.4.2
pluggy==0.6.0
py==1.5.3
pylint==1.8.4
pytest==3.5.1
pytest-bdd==2.21.0
python-dateutil==2.7.3
selenium==3.12.0
singledispatch==3.4.0.3
six==1.11.0
text-unidecode==1.2
virtualenv==15.2.0
wrapt==1.10.11
